Revert "Remove GUI interaction tests"
Previously, a change was merged to master prematurely,
removing all GUI interactions tests. That made the "tdm-nightly" job
to fail because the change in TDM related to the removal of these
interaction tests was not merged yet.

This change reverts the premature change, making the "tdm-nightly" job
to be able to succeed.

Change-Id: I118f3827e880d2db557f24fd9dbcc9bf50ca193f