mockup_travel: Revert default parameter removal
In a previous change the '<default>' element was removed accidentally
and ahead of time. It shouldn't have been removed within that change
since its logic is not yet removed and caused TDM nightly to fail.

This change brings back the removed element,
so that the TDM nightly job can succeed.

Change-Id: I279694d5744b1fab9fe584c107fb396a77a9653c