rasa_for_static_entities: Test entity synonyms
Before this change, there were no tests that validated that entity
synonyms generated by 'tala generate-rasa' worked all the way through
training and into TDM at run-time.

This change adds variations on an individual in
'rasa_for_static_entities', causing it to be generated as an entity
synonym in the training data for Rasa NLU. It also adds an interaction
test to validate that TDM can handle it at run-time.

Change-Id: Ieb3dce183c61ba1caefc0c5d26d2f25285343944