jenkinsfiledocker: Use TDM 'master'
Before this change, interaction tests in 'jenkinsfiledocker' used TDM
with the 'latest' Docker tag. This meant that the DDDs would not be
tested before future TDM changes were released, risking errors to
surface late.

This change ensures that TDM 'master' is used instead, which is the
first tag to be pushed when new TDM changes are merged to the master
branch, ensuring that errors are found early.

Change-Id: I0963ed815def54866af180814e9227a197b548b5