Rename datetime to datetime_ddd
Before this change, sometimes when running TDM tests from the
internal_ddds folder, TDM would crash saying that the 'datetime' module
did not have the expected methods.

This was however a misunderstanding, since the Python interpreter
mistakingly used the 'datetime' DDD where it had intended to use the
Python-builtin 'datetime' module. This happened since the
'internal_ddds' folder sometimes needed to appear on the PYTHONPATH.

Change-Id: I31c37b3d479b69090bb5168c7957cef7e26e48cf