mockup_travel: Remove misleading interaction test
This change removes the interaction test "sortal ambiguity for
recognized entities" since the functionality it covers doesn't seem to
be supported by TDM. The test seems to imply the following
functionality: If a user utterance contains a substring that matches two
entities of different sorts, and an answer move for only one of the
sorts is relevant in the dialogue context, then that answer move should
be integrated.

To the extent that this functionality is intended to be supported, the
implementation seems flawed. With the previous Python 2 version of TDM,
the removed test succeeded when run together with the other tests in the
file, but failed when run in isolation. With Python 3, the removed test
always fails, but succeeds if the device module returns the entities in
reverse order compared to today.

Presumably, when this kind of sortal ambiguity occurs today, TDM only
considers the first entity in its internal representation. In other
words, there seems to be no support for contextual disambiguation of
this kind.

Change-Id: Iacf50dceec4d2a46b427c0cfb668a5d6d9f7ec99