Make downdate conditions sharper
This change adapts the mockup_travel DDD to the changes in tala and TDM
which introduce a new way of specifying downdate conditions. The purpose
is to test the new features in TDM.

It also adds a new DDD, downdate_conditions, which is intended solely
for interaction testing of the new downdate conditions.

Change-Id: Ie088dcfedfb894732faeb5006a9dd7b1e880c958