Epic #1714: Intelligent feature question choice
This change adds support for testing the
"Intelligent feature question choice" feature,
which means that the system does not ask about
a feature if all alternatives have the same one.
E.g. if all phone directory hits are in Gothenburg,
the system does not ask about which city, but
continues to another feature question.

This change is related to the TDM change 1670:
http://gerrit.talkamatic.se:8080/c/tdm/+/1670

This change moves the old change with the same name
in the phone_directory ddd from tdm/ddds to
the internal_ddds submodule.

Change-Id: Ie61923ae83828d9aefb7e5ff854aa81d40062485