Epic #1714: Post-system-answer answer revision
This change adds support for testing the
"Post-system-answer answer revision" feature,
which means that the user can ask for another
answer after already receiving one from the system.
E.g. the user can ask for the number to another
person after receiving the number to one.

This change is related to the TDM change 1690:
http://gerrit.talkamatic.se:8080/c/tdm/+/1690

This change moves the old change with the same name
in the phone_directory ddd from tdm/ddds to
the internal_ddds submodule.

Change-Id: Ie8ffce4e1618e324593ee16510a026ee802e5007