Epic #1714: Providing number of alternatives
This change adds support for testing the
"Providing number of alternatives" feature,
which means that the number of hits is now
mentioned by the system before asking further
feature questions to narrow down the search
results.

It also adds changes to the interaction tests
that reflect this also in the semantic expressions
for testing where NLG is currently missing.

This change is related to the TDM change 1714:
http://gerrit.talkamatic.se:8080/c/tdm/+/1714

This change moves the old change with the same name
in the phone_directory ddd from tdm/ddds to
the internal_ddds submodule.

Change-Id: I1b39e469c478542d686e4f39248baf30681ce82c