Migrate interaction tests from TDM
Before this change, TDM contained a pytest-runnable test suite to build
and test the internal DDDs in tdm/ddds, which was used by the TDM
Jenkins jobs and could be run manually on a local machine.

An upcoming change will however merge the tdm and the internal_ddds
repositories, ensuring that both TDM and the DDDs are available to
Jenkins when running the tests. This allows the tests to be migrated to
the internal_ddds scope while still allowing them to be run manually and
by Jenkins.

This change migrates the tests here from the tdm repository, in
preparation for the merge with TDM.

Change-Id: I6267755a48ee76b8d9db549fb786a9376a47b9df