Remove icm:loadplan in top-level plans
This change updates the interaction tests to reflect the changes in
TDM with id: I94c63da2ebfb38cc811875456cf984c9c9532245

Background: icm:loadplan has not been verbalized for a long time, since
it was considered to be uttered a bit too often and not always in
appropriate contexts. Lately, we have started to verbalize it again.

Previously, the icm:loadplan ("Let's see.") was added when any action
plan asking a question was loaded, including the top plan. This lead
to users hearing the "Let's see" utterance at the start of any dialogue.
This was considered tedious.

The change makes sure that it icm:loadplan is not generated for top
plans.

Also: - move randomized interaction tests for domain_query to its own
   file
 - changing the name of the interaction test file for mockup_travel
   (from ending in ".semantic" to ".sem")
 - added a property Goal.type_ (to avoid collision with reserved word
   "type")
Change-Id: Idb41ddeac2169b6bbb1765e44239968b02f76e0f