Add temperature DDD
Previously, the testing of all "classic" TDM functionality was contained
in the DDD "mockup_travel". That DDD is a huge monster, and its tests
take forever to run.

This change adds a simple DDD for asking questions with one single
parameter, and is intended to be the start of splitting up the testing
done in "mockup_travel" into several, smaller test DDDs.

Change-Id: Iea2bb1895e73e3c4a09179680c615c9fc2b9bedf